+++ src/types/internal.ts
@@ -1,8 +1,10 @@
 import type { ServerResponse } from "http";
 import type { OpenAPIV3 } from "openapi-types";
-import type { HttpMiddleware, HttpPathValidator, HttpRequestHandler, IControllersOptions, IHttpController, IHttpServer, ImportValues, ParameterOptions } from ".";
+import type { HttpMiddleware, HttpPathValidator, HttpRequestHandler, IControllerMethodInfo, IControllersOptions, IHttpController, IHttpServer, ImportValues, ParameterOptions } from ".";
 
 export type Constructor<T extends any = any> = (new (...args: any[]) => T);
+
+export type Func = (...args: any[]) => any;
 
 export type GetterFunc<TValue extends any = any> = () => TValue;
 
@@ -49,6 +51,7 @@ export interface IInitControllerMethodActionContext {
     globalOptions: Nilable<IControllersOptions>;
     method: Function;
     relativeFilePath: string;
+    resolveInfo: ResolveControllerMethodInfo;
     server: IHttpServer;
 }
 
@@ -115,6 +118,13 @@ export interface IPrepareControllerMethodActionContext {
     server: IHttpServer;
 }
 
+export interface IRequestHandlerContext {
+    end: (response: ServerResponse) => void;
+    handler: HttpRequestHandler;
+    isPathValid: HttpPathValidator;
+    middlewares?: Nilable<HttpMiddleware[]>;
+}
+
 export interface ISwaggerMethodInfo {
     doc: OpenAPIV3.PathItemObject;
     method: Function;
@@ -135,9 +145,4 @@ export type Optional<T extends any = any> = T | undefined;
 // s. https://stackoverflow.com/questions/43159887/make-a-single-property-optional-in-typescript
 export type PartialBy<T, TKey extends keyof T> = Omit<T, TKey> & Partial<Pick<T, TKey>>;
 
-export interface IRequestHandlerContext {
-    end: (response: ServerResponse) => void;
-    handler: HttpRequestHandler;
-    isPathValid: HttpPathValidator;
-    middlewares?: Nilable<HttpMiddleware[]>;
-}
+export type ResolveControllerMethodInfo = (info: IControllerMethodInfo) => any;